Move methods to form component
