+++ src/Components/Form/Form.js
@@ -1,7 +1,46 @@
 import './Form.css'
+import { useState } from 'react'
+import { postQuestion } from '../../apiPrompt'
 import PropTypes from 'prop-types'
 
-const Form = ({ updateQuestion, submitQuestion, value }) => {
+const Form = ({ conversation, setConversation, setIsLoading }) => {
+  const [question, setQuestion] = useState('')
+
+  const updateQuestion = (e) => {
+    setQuestion(e.target.value)
+  }
+
+  const submitQuestion = (e) => {
+    e.preventDefault(e)
+    validate()
+    postQuestion(question)
+      .then((response) => {
+        if (!response.ok) {
+          alert('Uh oh. Please try another question.')
+        } else {
+          return response.json()
+        }
+      })
+      .then((answer) => {
+        const questionAnswer = {
+          id: Date.now(),
+          you: question,
+          oracle: answer.choices[0].text
+        }
+        setConversation((prevState) => [questionAnswer, ...conversation])
+        setIsLoading(false)
+        clearInput()
+      })
+  }
+
+  const clearInput = () => {
+    setQuestion('')
+  }
+
+  const validate = () => {
+    !question ? alert('Please enter a question.') : setIsLoading(true)
+  }
+
   return (
     <form className='question-form fade-in-top' onSubmit={submitQuestion}>
       <label htmlFor='questionBox' className='label'>
@@ -10,7 +49,7 @@ const Form = ({ updateQuestion, submitQuestion, value }) => {
       <input
         type='text'
         id='questionBox'
-        value={value}
+        value={question}
         className='question-box'
         onChange={updateQuestion}
         autoFocus
@@ -29,7 +68,7 @@ const Form = ({ updateQuestion, submitQuestion, value }) => {
 export default Form
 
 Form.propTypes = {
-  updateQuestion: PropTypes.func.isRequired,
-  submitQuestion: PropTypes.func.isRequired,
-  value: PropTypes.string.isRequired
+  conversation: PropTypes.arrayOf(PropTypes.object).isRequired,
+  setConversation: PropTypes.func.isRequired,
+  setIsLoading: PropTypes.func.isRequired
 }
